Move GoTask to its own file
* Move the GoTask to its own file.
* Create a GoTask interface so the task can update its caller (I have no idea if this is right)
