+++ app/src/main/java/lunchandlearn/android/com/hermes/interfaces/GoTaskListener.kt
@@ -1,0 +1,6 @@
+package lunchandlearn.android.com.hermes.interfaces
+
+interface GoTaskListener {
+    fun showProgressBar(show: Boolean)
+    fun onTaskCompleted(canceled: Boolean)
+}
